% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_variables.R
\name{screen.glmnet.fix}
\alias{screen.glmnet.fix}
\title{Lasso screener selects pre-specified variables}
\usage{
screen.glmnet.fix(Y, X, family, alpha = 1, minscreen = 2, nVar = 10,
  nfolds = 10, nlambda = 100, fixed.var.index = var.index, ...)
}
\arguments{
\item{Y}{outcome variable (specified in \code{SuperLearner()})}

\item{X}{data frame}

\item{nVar}{number of non-zero variables to be selected}

\item{var.index}{indices of variables to always be included by the screener}
}
\description{
Lasso screener for \code{SuperLearner()} that always retains specified variables and passes
approximately \code{nVar} variables to \code{SuperLearner()}. When the number of non-zero
coefficients exceeds \code{nVar}, a larger value of the regularization parameter \code{lambda}
is chosen to select a smaller set of variables that excludes the ties.
}
\examples{
If you do not know the indices of the variables you always want to include, 
  you can get them from the variable name, where newdat is the dataframe: 
  var.index <- c(which(colnames(newdat)=="tcls14"), 
                 which(colnames(newdat)=="tcls251"))
}
\seealso{
See \code{\link[glmnet]{glmnet}} for additional details on implementing lasso
}
\keyword{EvilInsurers}
\keyword{SuperLearner}

