% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_variables.R
\name{screen.rf.exact}
\alias{screen.rf.exact}
\title{Random forest screener limits selected variables and selects pre-specified variables}
\usage{
screen.rf.exact(Y, X, family, nVar = 10, nFix = 3,
  fixed.var.index = var.index, ntree = 500, mtry = ifelse(family$family ==
  "gaussian", floor(sqrt(ncol(X))), max(floor(ncol(X)/3), 1)),
  nodesize = ifelse(family$family == "gaussian", 5, 1), ...)
}
\arguments{
\item{Y}{outcome variable (specified in SuperLearner())}

\item{X}{data frame}

\item{nVar}{number of variables for the screener to select}

\item{nFix}{number of individual variables that are alaways passed to SuperLearner()}

\item{var.index}{indices of variables to always be included by the screener}
}
\description{
Random forest screener for \code{\link[SuperLearner]{SuperLearner()}} that selects specified individual variables
and specified overall number of variables.
}
\details{
This function can be pretty slow, because currently it
operates by searching the rankings for the user selected ("fixed") variables. If the fixed variables
are included in the top \code{nVar} then it does not change anything. If the fixed variables are
not included in the top \code{nVar}, then it selects a subset of top \code{nVar}; e.g., the overall number
of variables to select is 10 and 2 of the fixed variables are outside the top 10, it will select the
top 8, and convert the 2 fixed variables outside the top 10 to be \code{TRUE}.
}
\section{Super Learner}{

See \code{SuperLearner()} documentation for information on additional arguments and
instructions on implementing \code{SuperLearner()}.
}
\examples{
If you do not know the indices of the variables you always want to include, 
 you can get them from the variable name, where newdat is the dataframe name:
 
 var.index <- c(which(colnames(newdat)=="sex"), which(colnames(newdat)=="age"), 
               which(colnames(newdat)=="emp_active")) 

}
\seealso{
\code{\link{screen.glmnet.fix}} for lasso screener
}
\keyword{EvilInsurers}
\keyword{SuperLearner}

