% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_variables.R
\name{screen.rf.fuzzy}
\alias{screen.rf.fuzzy}
\title{Random forest screener selects pre-specified variables}
\usage{
screen.rf.fuzzy(Y, X, family, nVar = 10, ntree = 500,
  mtry = ifelse(family$family == "gaussian", floor(sqrt(ncol(X))),
  max(floor(ncol(X)/3), 1)), nodesize = ifelse(family$family == "gaussian", 5,
  1), ...)
}
\arguments{
\item{Y}{outcome variable (specified in \code{SuperLearner()})}

\item{X}{data frame}

\item{nVar}{number of variables for the screener to select}

\item{var.index}{indices of variables to always be included by the screener}
}
\description{
Random forest screener for \code{\link[SuperLearner]{SuperLearner()}} that selects
user specified variables in addition to variables chosen data-adaptively
}
\details{
If you do not care about the exact number of variables the screener chooses, use this function
rather than \code{screen.rf.fix.exact}. This function is faster, but will not necessarily return 
exactly \code{nVar} variables to \code{SuperLearner()}. \code{screen.rf.fuzzy} selects the top
\code{nVar} variables, and then also makes sure the user specified variables are also passed
to \code{SuperLearner()}. If the user specified variables are in the top \code{nVar} variables,
then \code{nVar} variables will be passed to \code{SuperLearner()}. If any of the user 
specified variables are outside the top \code{nVar} variables, then more than \code{nVar}
variables will be passed to \code{SuperLearner()}.
}
\section{Super Learner}{

See \code{SuperLearner()} documentation for information on additional arguments and
instructions on implementing \code{SuperLearner()}.
}
\examples{
If you do not know the indices of the variables you always want to include, 
 you can get them from the variable name, where newdat is the dataframe name:
 
 var.index <- c(which(colnames(newdat)=="sex"), which(colnames(newdat)=="age"))
}
\seealso{
\code{\link{screen.glmnet.fix}} for lasso screener, \code{\link{screen.rf.exact}}
for exact random forest screener.
}
\keyword{EvilInsurers}
\keyword{SuperLearner}

