% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_variables.R
\name{tgrp.fun}
\alias{tgrp.fun}
\title{Examine only therapeutic groups from Evil Insurers Marketscan sample}
\usage{
tgrp.fun(X, ...)
}
\arguments{
\item{X}{data frame columns (should be specified when calling \code{SuperLearner()})}
}
\description{
This function allows you to screen out the therapeutic class variables when running 
\code{SuperLearner()}. It relies on the therapeutic class variables all beginning with
the prefix \code{tcls}. I should probably make this a more general function that allows
the user to specify whatever variable prefix they want.
}
\examples{
Cannot think of a good example here -- pretty much the one setting for using it

}
\keyword{EvilInsurers}
\keyword{SuperLearner}

